New statistic for counting nr cluster.
Fix for dnaio file reading where error appears if fastq read sequence and qualities strings are of different lengths. This can appear in our code as the sequence is replaced by the error corrected sequence in this script. Fix replaces qualities string with string of same length and sequence but containing only 'G'.